# core/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: NeuroGate API
  description: |
    Enterprise AI Gateway API providing intelligent LLM routing, PII protection,
    semantic caching, and multi-provider orchestration.

    ## Features
    - **Multi-Provider Routing**: Intelligent routing across OpenAI, Anthropic, Google, AWS Bedrock, and Azure
    - **PII Protection**: Automatic detection and redaction of sensitive data (NeuroGuard)
    - **Semantic Caching**: 4-tier caching with semantic similarity matching
    - **Cost Tracking**: Real-time usage and cost analytics
    - **Agent Operations**: Trace and monitor AI agent workflows

    ## Authentication
    Use API key authentication via the `Authorization` header:
    ```
    Authorization: Bearer your-api-key
    ```

  version: 1.0.0
  contact:
    name: NeuroGate Team
    url: https://neurogate.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local Development
  - url: https://api.neurogate.dev
    description: Production
...
paths:
...
  /v1/models/{modelId}:
    get:
      tags: [Models]
      summary: Get model details
      description: Get detailed information about a specific model
      operationId: getModel
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfo'
        '404':
          description: Model not found
...
  /api/v1/analytics/costs/user/{userId}:
    get:
      tags: [Analytics]
      summary: Get user cost report
      operationId: getUserCosts
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Cost report retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostReport'
  /api/v1/analytics/costs/team/{teamId}:
    get:
      tags: [Analytics]
      summary: Get team cost report
      operationId: getTeamCosts
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Team cost report retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamCostReport'

  # ==================== NeuroGuard API ====================
...
  /api/hive/consensus:
    post:
      tags: [Consensus]
      summary: Reach consensus
      operationId: reachConsensus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Consensus reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsensusResult'

  # ==================== AgentOps API ====================
...
  /v1/agent/memory:
    post:
      tags: [Memory]
      summary: Store memory
      operationId: storeMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreMemoryRequest'
      responses:
        '200':
          description: Memory stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  memory_id:
                    type: string
    get:
      tags: [Memory]
      summary: Retrieve memory
      operationId: retrieveMemory
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: Memories retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  # ==================== Flywheel API ====================
  /v1/flywheel/feedback:
    post:
      tags: [Flywheel]
      summary: Submit feedback
      operationId: submitFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '200':
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoldenInteraction'
...
  /v1/flywheel/export:
    get:
      tags: [Flywheel]
      summary: Export dataset
      operationId: exportDataset
      parameters:
        - name: model
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Dataset exported
          content:
            application/jsonl:
              schema:
                type: string

  # ==================== Debugger API ====================
...
components:
  schemas:
...
    ChatRequest:
      type: object
      required:
        - messages
      properties:
        model:
          type: string
          default: gpt-4o
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        temperature:
          type: number
          default: 0.7
        max_tokens:
          type: integer
        stream:
          type: boolean
          default: false
        tools:
          type: array
          items:
            type: object
        tool_choice:
          type: string
...
    Message:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
        name:
          type: string
        tool_calls:
          type: array
          items:
            type: object
...
    ModelInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        input_cost_per_1k:
          type: number
        output_cost_per_1k:
          type: number
        context_window:
          type: integer
        max_output_tokens:
          type: integer
        capabilities:
          type: array
          items:
            type: string
        available:
          type: boolean
...
    CostReport:
      type: object
      properties:
        user_id:
          type: string
        period:
          type: string
        total_cost:
          type: number
        total_requests:
          type: integer
        cache_hits:
          type: integer
        cache_hit_rate:
          type: number
        provider_breakdown:
          type: object
        model_breakdown:
          type: object
    TeamCostReport:
      type: object
      properties:
        team_id:
          type: string
        period:
          type: string
        total_cost:
          type: number
        user_breakdown:
          type: object

    # ==================== Security Schemas ====================
...
    ConsensusResult:
      type: object
      properties:
        consensus:
          type: string
        responses:
          type: array
          items:
            type: object
        confidence:
          type: number

    # ==================== Trace Schemas ====================
...
    StoreMemoryRequest:
      type: object
      properties:
        content:
          type: string
        metadata:
          type: object

    # ==================== Flywheel Schemas ====================
    FeedbackRequest:
      type: object
      properties:
        trace_id:
          type: string
        session_id:
          type: string
        rating:
          type: integer
        feedback_text:
          type: string
        feedback_type:
          type: string
        messages:
          type: array
          items:
            type: object
        response:
          type: string
        model:
          type: string
        corrected_output:
          type: string
    GoldenInteraction:
      type: object
      properties:
        id:
          type: string
        trace_id:
          type: string
        timestamp:
          type: string
          format: date-time
        rating:
          type: integer
        feedback_text:
          type: string

    # ==================== Debugger Schemas ====================
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []